requirements.txt
# Retailer Scraping Engine — Phase 1
# Python 3.10+

playwright>=1.40.0
httpx>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0

flask>=3.0.0
